Multiple Updates to Player and Traps
- Removed Friction from Traps
- Separated Proximity Detector
- Changed Player Speed
- Changed Activation Distance of Sliding Block Trap
