+++ Assets/Scenes/Traps.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d645ae34f0cb36346b7da8eede3247b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
