+++ Assets/Scenes/Traps/Sliding Block Trap.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5bcdfedc1458fab41878565d04c4b528
+guid: 15a5112e7f1d2834bbc45300014b09db
 DefaultImporter:
   externalObjects: {}
   userData:
