--- Assets/Scripts/MoveSprite.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b37c7d9213aab1a42bc58c2c0e8b7273
