+++ Assets/Scripts/MoveTrap.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 36384c21ac78af34d96c45689eb72cc4
