+++ Assets/Scripts/ProximityDetector.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f10cb97b72ce6284d9509fbf9a184f9a
